+++ NEWS.md
@@ -1,3 +1,9 @@
+# naijR 0.4.5
+## Enhancements:
+
+* Effectively handle mobile numbers that have common separators in them, namely whitespace, '-' or '.'.
+* Repair mobile numbers where poor data entry interchanges zeros (`0`s) with the letter `O` (works for both upper and lower case).
+
 # naijR 0.4.4
 ## Bug fix: